build: move tests and fixtures to the root dir.
