+++ packages/animeflv/__tests__/url.test.ts
@@ -1,4 +1,4 @@
-import { animeSlug, browseAnime, browseFilter, verSlugEpisode } from '../index.js';
+import { animeSlug, browseAnime, browseFilter, verSlugEpisode } from '../src/index.js';
 
 describe('browseAnime()', () => {
   it('returns with query q', () => {
