+++ packages/animeflv/tsconfig.json
@@ -1,9 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist"
+    "rootDir": "."
   },
-  "include": ["src"],
+  "include": ["src", "__tests__", "fixtures"],
   "exclude": ["node_modules"]
 }
